Major: mkphotodiode -- working prototype
Triggers on SampleCommand and takes in photodiode input on arduino analog pin. In preliminary tests, browser can receive samples every ~300 microseconds from arduino. So can communicate at 1 kilohertz speeds and can send many values over at a time.
Plot updates are a rate-limiting step and are best done in between triggers so as not to slow down acquisition.
Plotting is limited in this prototype, so need to work on that and sending over multiple values.
Current version was throttled to run at 1kHz and sends over 1 value at a time.