Added display priming
Display priming not as necessary for useCanvas2D mode. Can help with reliable first frame onset timing when using webgl canvas for 3D rendering. Priming the display is somewhat like double buffering. Although this seems to allow the sample frame to come on in the same, close in time to when it was issued, it does add a tens of milliseconds lag in bringing up that first frame. One lesson learned is that mirror displays produces better, more reliable onset timing of first frame and reduces frame drops of later frames.

Added check to make sure BigQuery save calls are made even if on CURRTRIAL.num =0, it doesn't register for some reason.