MAJOR: mkphotodiode upgrades
mkphotodiode compares the current trace to the previous one. Displays comparison of onset and duration of the two traces as well as sampling rate states on the arduino sampling and webusb pipe sample intervals. Runs at 1.5kHz. Sampling interval averages 0.6-0.7ms but with shorter and longer intervals. Implemented interpolation and capability to do correlation between two waveforms.

Arduino code modified to send timestamps. This gives more precision on sample timings and is done relative to last sample on trigger. Effectively the sample command serves as a syncing signal -- upon each one all clocks are resynchronized. So mkphotodiode javascript will add the arduino time to when it last received a sample command over the webusb pipe (as soon as arduino sees sample command go high, it calls serial.flush to get it to mkphotodiode).

This is a near final version for using mkphotodiode as a testing unit. Remaining to be implemented for full integration with mkturk: syncing with mkturk trial count, saving to local disk, saving to bigquery