explicit passing of images & clip# to displayTrial
So that priming images does not interfere with actual display, needed to stop using a global variable for storing images and sequenceclip#. Now displayTrial & displayTrial_prime explicitly take in the images for that trial and clip# and pass to render3d and render2d.
Fixed indexing bug in retrieving trialqueue samplefilename.