mkeye: add line plot of x,y trajectories
Example for using charts.js to make line plots with two datasets (incorrect & correct trials).

EVENTS.timeseries.EffectorXY touch data is stored in screen coordinates (with y-axis inverted) since that is the way eye data are stored. The rtdb code inverts y-axis and subtracts canvas offsets in the hold listener, but this isn't what we want for data storage.