Send displayed objects + target boxes + effector state on every frame flip
Unifies sending of displayed assets' physical extent in 2D pixels, the associated target windows, and the effector state. This is now done on every frame flip so that have an updated state on each screen refresh which is usually @ 60Hz. This seems to be adequate for tracking most behavioral operations.

Note that the eyetracker or touchscreen can sample at a faster rate than 60Hz and these raw data are still stored using logEvents timeseries. These extra samples within each frame can be integrated using a median filter to get a better estimate of the effector state with each frame flip. For example, the eyetracker signal is median filtered over four samples before being sent to the touchhold_listener.
Regardless, for closed-loop behavior control and for user-facing display in realtime, 60Hz gives a fixed clock for how often mkturk rendering can be updated and how often viewers like mkeye would be able to display as both are bounded by the display refresh rate.