mkturk_three.js accommodates toggling textures off
Default is to render any textures included in the glb file. However if texture = "false" for that object in the scene file, then the specified material in the scene file will be used. Prior behavior was that this specified material was used for any mesh named "Base", but now it can be toggled for any mesh if user specified texture = false in the scene file. All child meshes will get the default material regardless of whether they had a material or not in the glb file when texture = false.

New variable "texture" is no vectorizable. For now, the mesh + texture are treated as fixed primitives for rendering. If you want a different mesh+texture combo, then add another object to a scene file and toggle it visible (e.g., if want to add/remove texture). This new object whether with or without texture is treated as a new identity in the scene.
Under the hood, mkturk_threejs preloads the mesh and texture separately from frame by frame updates to the scene. In the future, could revisit this if we want to change textures frame by frame. Morphing the mesh is an example where we morph the mesh and effectively allow an identity change in the movie. One thing to check if we change textures in realtime is whether this significantly slows down the rendering.