Restore separate rendering and filtering of objects & background cubemap
Since the move to cubemap instead of images, render3D in mkturk_screenfunctions had dropped support of separate filtering of objects and background. This is now restored by allowing for two rendering styles:

1 - no filtering - simultaneously render all objects and cube map in one shot

2 - object and/or image filter specified  - render the background/filter/transfer then in a second render do foreground objects, apply object specific filter then transfer.

Method 1 is just the mode being used up to this point.

Method 2 gives the flexibility needed to toggle filters on objects and background separately. However, this could be slower since it is two render calls. Also, if the camera is outside the cubemap, then it will create a non-sensical image where the objects are overlaid on the cubemap even though they should be occluded by the front wall (assuming they reside within the room or cubemap)

Also added filtering functionality for 2D image rendering mode (threejs is not used if all sample & test scenefiles are image only)