MAJOR: Multidevice usb using BroadcastChannel
Multiple tabs each will claim one usb device and broadcast which device index they claimed. The broadcast is done every couple seconds, so it's effectively like renewing a lease on the device.
If the tab doesn't broadcast, then it is assumed that that particular device index is now free.

One major drawback of this endeavor inherent to the WebUSB api is that when a device is disconnected, it is not known which one was disconnected as the API does not report that, so all tabs will cease their usb. What this means is that in the event of a single disconnection on that computer/tablet/phone, all connections for all tabs should be re-established.

BroadcastChannel worked out well here because it can be used within the usb web worker and is not limited to strings whereas localstorage can only be called from the main thread and sends strings.