MAJOR: upgrades for block design, more triggers
Block design treated as a first class citizen.
(1) Added new trialseries variable BlockNum to keep track of blocks in addition to trials in data file -- see EVENTS['trialseries']['BlockNum']

(2) Arduino now outputs two duplicate trial triggers and two duplicate block triggers in addition to the sample command line which contains the filecode and sample sync signal. In the future the second trial and block trigger lines can encode other properties, but are duplicates for now. Also, in the future, the logic of what constitutes a "block" can be user-controlled so that have other independent timing signals in addition to the trial structure.

In mkturk_trialqueue_scene.js, made the decision to unify how blocks are done across specifying nstim versus nmilliseconds. Now block gating is done at the end of get_next_trial. In other words, stim are generated from all bags and then when get trial, see if match bag block requirement -- if not, then discard. This ensures priming of all bags upon loading and simplifies the code for block type logic. Now index.js handles the incrementing of the block number, so this is now done outside of trialqueue_scene which is more transparent than burying in trialqueue_scene.

Minor: fixed issue that filecode was not sent when auto-switching to new task if params changed. Filecode was only sent when user left practice mode, but now filecode is sent any time the task parameters are reloaded and consequently a new data file will be created, requiring a new filecode send.